client initialization decorator breaking changes
